full (responsive) frontend for the entire 4-full-mvc-framework | fixes

+++ 4-full-mvc-framework/views/note/index.php
@@ -1,23 +1,32 @@
 <div class="content">
     
-    <h1>My Notes</h1>
-    <h3>This screen shows/handles just the user's notes</h3>
+    <h1>Create new note</h1>
+    <h3>"Notes" are just an example of how to create, edit and delete things.</h3>
 
     <form method="post" action="<?php echo URL;?>note/create">
         <label>Text of new note: </label><input type="text" name="note_text" />
         <input type="submit" value='Create this note' />
     </form>
 
-    <hr />
+    <h1 style="margin-top: 50px;">List of your notes</h1>
 
     <table>
     <?php
-        foreach($this->notes as $key => $value) {
-            echo '<tr>';
-            echo '<td>' . $value->note_text . '</td>';
-            echo '<td><a href="'. URL . 'note/edit/' . $value->note_id.'">Edit</a></td>';
-            echo '<td><a href="'. URL . 'note/delete/' . $value->note_id.'">Delete</a></td>';
-            echo '</tr>';
+    
+        if ($this->notes) {
+            
+            foreach($this->notes as $key => $value) {
+                echo '<tr>';
+                echo '<td>' . $value->note_text . '</td>';
+                echo '<td><a href="'. URL . 'note/edit/' . $value->note_id.'">Edit</a></td>';
+                echo '<td><a href="'. URL . 'note/delete/' . $value->note_id.'">Delete</a></td>';
+                echo '</tr>';
+            }            
+            
+        } else {
+    
+            echo 'No notes yet. Create one !';
+        
         }
     ?>
     </table>
